Remove Options Class
According WP Review, we cannot store data in variables without sanitizing it, even if the data is sanitized when it is saved origianlly, and again according WP Review, it is wrong to escape when assigning data to a variable. While I have trouble understanding this, having it never seen nor reead anywhere and seeing even core does not folloow this rule, I have no choice but to comply, and with it, the class becomes obsolete. I will figure out WHY this should be done as asked for ... and update my own knowledge :)